Implementation Plan - Align Win32Emulator Shellcode Loading & Execution This plan details the design and proposed changes to fully port the shellcode loading (load_shellcode) and execution (run_shellcode) logic in Win32Emulator from Python to C++, and to refactor thread stack base and commit types to uint64_t to prevent truncations on x64 architectures.
User Review Required
We are updating the following components:

Thread stack type safety: Refactoring Thread's stack_base_ and stack_commit_ fields from int to uint64_t.
Win32Emulator::load_shellcode: Transitioning from raw mapping as PERM_MEM_RW to high-fidelity wrapping via ShellcodeLoader (mapping as executable PERM_MEM_RWX), adding proper file-read fallback logic when data is empty, supporting filename override, computing SHA-256 hex string using picosha2, and populating profiler input metadata.
Win32Emulator::run_shellcode: Aligned to validate the shellcode address, configure the stack, allocate process and TEB contexts, map 4 dummy stack arguments of size 1024 bytes, set the ECX register value to 1024, and bind process/thread environments.

+++ secpp/windows/objman.cpp
@@ -490,7 +490,7 @@ IoStackLocation Irp::get_curr_stack_loc() {
 // Thread
 // 
 
-Thread::Thread(void* emu, int stack_base, int stack_commit)
+Thread::Thread(void* emu, uint64_t stack_base, uint64_t stack_commit)
     : KernelObject(emu),
       ctx_(nullptr), modified_pc_(false), teb_(nullptr),
       suspend_count_(0), token_(emu), last_error_(0),
@@ -563,9 +563,9 @@ void Thread::set_context(void* ctx) {
     this->ctx_ = ctx;
 }
 
-void Thread::init_teb(int teb_addr, int peb_addr) {
+void Thread::init_teb(uint64_t teb_addr, uint64_t peb_addr) {
     if (!this->teb_) {
-        this->teb_ = std::make_shared<TEB>(emu_, teb_addr);
+        this->teb_ = std::make_shared<TEB>(emu_, static_cast<int>(teb_addr));
     }
 
     auto* teb_struct = static_cast<speakeasy::defs::nt::TEB*>(this->teb_->get_object());
